OpenMP for the galaxy benchmark (#246)
This adds OpenMP to:
* load balancing
* godunov solver AMR and source term parts
* gas gravity parts

It also adds the galaxy setup as a test.

---------

+++ tests/poisson/cosmo-amr/before-test.sh
@@ -1,10 +1,15 @@
 if [ ! -d "cosmo_128" ]; then
-    echo "cosmo_128 directory does not exist, dowloading it..."
-    wget https://ramses.cnrs.fr/wp-content/uploads/2025/01/cosmo_128.zip
-    unzip cosmo_128.zip
-    rm cosmo_128.zip
+    if [ ! -d "../cosmo/cosmo_128" ]; then
+        mv ../cosmo/cosmo_128 ./
+        echo "cosmo_128 directory moved from ../cosmo/cosmo_128."
+    else
+        echo "cosmo_128 directory does not exist, dowloading it..."
+        wget --timeout=10 --tries=3 --no-check-certificate https://ramses.cnrs.fr/wp-content/uploads/2025/01/cosmo_128.zip
+        unzip cosmo_128.zip
+        rm cosmo_128.zip
 
-    echo "cosmo_128 directory downloaded."
+        echo "cosmo_128 directory downloaded."
+    fi
 else
     echo "cosmo_128 directory already exists."
 fi
